Add support for 'generic' quantities
Potentially useful for cases where the correct units type is unclear, or if there fundamentally is no units type. For example, a logo may have no fundamental units - it may be displayed in pixels or printed out in meters in various different scales but has no inherent size of its own. As a result, it may make sense to define the logo in generic 'Units' and then convert to different actual units as needed.