Remove 'Fractional'/'Whole' type aliases
Not totally happy with the names, and they add another layer of indirection/abstraction in type signatures (which can make compiler error messages more confusing). So leave them out for now, and if needed something like them can be added in the future without triggering a major version bump (as opposed to leaving them in for now and removing/renaming them later, which *would* be a breaking change.)

Note that Quantity.toFractional is no longer required since it can be replaced by 'Quantity.map toFloat'.