Add Molarity module
See #6.

An SI unit, mole per cubic meter (mol/m3), `Molarity` is an amount of
substance divided by the volume of the mixture. Values stored in number
of mole per cubic meter. All `*PerLiter` definitions are defined in
terms of `decimolesPerLiter`. Why? This was done because a liter equal
to a cubic decimeter, and `mole per cubic meter` is a `Rate`. This may
make the various `*PerLiter` definition appear off by a factor of 10
because the definitions are in terms of decimoles per cubic decimeter
(same as `decimolesPerLiter`).